requirements.txt
Django~=2.1.2
django_redis~=4.9.0
django-import-export~=1.1.0
django-extensions~=2.1.3
django-colorfield~=0.1.15
django-cache-machine~=1.0.0
PuLP~=1.6.8
psycopg2-binary~=2.7.5
requests~=2.19.1
numpy~=1.15.2
pyclustering~=0.8.1
channels~=2.1.3
channels-redis~=2.3.1
python-memcached~=1.59
isoweek~=1.3.3
django-ical
